network btn consolidation
